requirements.txt
amqp==2.5.1
billiard==3.6.1.0
cachetools==3.1.1
celery==4.3.0
Django==2.2.5
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
vine==1.3.0
zipp==0.6.0
